# Conflicts:
#	LocalVenue/Services/ShowService.cs

+++ LocalVenue/Controllers/RequestModels/TicketRequest.cs
@@ -4,7 +4,7 @@ namespace LocalVenue.RequestModels;
 
 public class TicketRequest
 {
-    public int? TicketId { get; set; }
+    public long? TicketId { get; set; }
     public long ShowId { get; set; }
 
     public long SeatId { get; set; }
